Load JSON configs to global var scenesMap
Move respective functions to manager as there's no
parsing involved.
