+++ scene/manager.go
@@ -5,8 +5,11 @@
 package scene
 
 import (
+	"encoding/json"
 	"io/ioutil"
 	"regexp"
+
+	"github.com/3ter/iMagine/fileio"
 )
 
 // ScenesDir publishes the directory where its files are stored
@@ -14,6 +17,23 @@ const ScenesDir = `../scene/`
 
 // scenesMap maps scene identifiers (e.g. 'Beach') to their respective scene object
 var scenesMap map[string]*Scene
+
+// MapConfig contains key/value-pairs for a scene that are intended to save
+// * which scenes are adjacent to the current one
+// * the state of the scene
+type MapConfig struct {
+	North string
+	East  string
+	South string
+	West  string
+	Look  string
+}
+
+func (s *Scene) loadMapConfig(filename string) {
+	jsonBytes := fileio.LoadFileToBytes(filename)
+
+	json.Unmarshal(jsonBytes, &s.mapConfig)
+}
 
 func loadFilesToSceneMap() {
 	scenesMap = make(map[string]*Scene)
@@ -25,19 +45,22 @@ func loadFilesToSceneMap() {
 	for _, sceneFile := range sceneFileSlice {
 		sceneFileFilter := regexp.MustCompile(`^scene(\w+)\.(md|json)$`)
 
-		FileMatchSlice := sceneFileFilter.FindStringSubmatch(sceneFile.Name())
-		if len(FileMatchSlice) == 3 {
-			FileScene := FileMatchSlice[1]
-			FileExtension := FileMatchSlice[2]
+		fileMatchSlice := sceneFileFilter.FindStringSubmatch(sceneFile.Name())
+		if len(fileMatchSlice) == 3 {
+			filePath := ScenesDir + fileMatchSlice[0]
+			fileScene := fileMatchSlice[1]
+			fileExtension := fileMatchSlice[2]
 
-			if scenesMap[FileScene] == nil {
-				scenesMap[FileScene] = getSceneObjectWithDefaults()
+			if scenesMap[fileScene] == nil {
+				scenesMap[fileScene] = getSceneObjectWithDefaults()
 			}
 
-			if FileExtension == `md` {
-				scenesMap[FileScene].script.filePath = ScenesDir + FileScene + `.md`
-			} else if FileExtension == `json` {
-				scenesMap[FileScene].mapConfigPath = ScenesDir + FileScene + `.json`
+			if fileExtension == `md` {
+				scenesMap[fileScene].script.filePath = filePath
+				scenesMap[fileScene].script.fileContent = fileio.LoadFileToString(filePath)
+			} else if fileExtension == `json` {
+				scenesMap[fileScene].mapConfigPath = filePath
+				scenesMap[fileScene].loadMapConfig(filePath)
 			}
 		}
 	}
